En esta confirmacion me enfoque en varios puntos, los cuales fueron los siguientes: A) Que cada dia que pase se registre y se pueda visualizar dicho cambio en el timeEdit de la GUI principal B) Que todos los cuadros de dialogo tuvieran sus respuestas y comunicado en idioma español C) Poner una excepccion en el caso de que se quiera cargar una cancion que ya fue cargada o mejor dicho cuyo nombre ya esta registrado D) Poner una excepccion en el script 'notificadorAlarmas.py' con la finalidad de que si quiere tocar una cancion que ya no existe o no hay cancion configurada para cancion entonces que no colapse el programa y que muestre la leyenda de 'sin musica' E) Me enfoque a realizar los audios respectivos de los temas de: HORA_DESPERTAR, HORA_DORMIR, HORA_TAREAS los cuales son de mi autoria ya que yo invente el dialogo y modifique la tonalidad de voz con ayuda de audocity y finalmente lo exporte en audio con formato '.wav' F) Me enfoque a desarrollar la ventana que mostrara mis datos por si alguien gustara contactarme o ver mi repositorio o perfil de likedin
Todos lo mencionado eran pendientes por realizar que finalmente fueron
terminados, sin embargo a la hora de correr la aplicacion en la rasberry
pi o mejor dicho en un el sistema operativo 'Debian' hubo  bastantes
complicaciones:

A)Los 'tabWidget' de la GUI principal no se veian bien, y se veian
amontodos, lo cual en windows 10 ocurria eso
B)A la hora de usar la funcion de pygame 'mixer.Sound()' con la
finalidad de reproducir los audios, me di cuenta que los archivo '.mp3'
no los corria correctamente, sin embargo si ejecutaba los que tenian
extencion '.wav'
C)En la aplicacion de 'DEBERES' precisamente en el script de
'SeccionNotas.py' me percate que a la hora de marca una nota como
realizada y por consecuente esta se tenia que eliminar de la vista,esto
me ocacionaba error, cosa que en windows no, despues de segui
minusiosamente el codigo  no pude encontrar el problema raiz, pero si
una solucion a corto plazo, lo cual fue almacenar las referencias de los
punteros de la widgets que representaban cada item de cada nota de  las notas
que se creaban y solo las podia eliminarlas de la lista dichas
referencias hasta que se decidiera eliminar otra, de esa forma funciono
eso, cabe mencionar que probe varias soluciones y delimite bastante
el problema pero parecia no tener solucion, y tambien es importante
mencionar que la solucion que implemente funciona de manera correcta
y ya no da errores y sobre todo casi no demanda recursos extras de
los que ya se demandaban anteriormente.

Una vez identificados esos problemas me vi en la necesidad de generar
las soluciones y eso fue lo que tambien realice en esta confirmacion.

Es importante mencionar que el trabajo realizado en esta confirmacion
ya permite almenos correr el programa en: 'windows 10' y 'rasbian'

+++ GUI/CUERPO/LOGICA/ALARMA/reproductorSonidosAlarmas.py
@@ -65,8 +65,19 @@ class ReproductorSonidosAlarmas(QtCore.QObject):
             print("Tamano cancion:",tamano)
             if tamano<self.MAX_DURACION:
                 #Copiando cancion...
-                shutil.copyfile(origenArchivoCopiar,destinoArchivoCopiar)
-                self.senal_cancionAgregada.emit(soloNombreCancion)
+                if os.path.exists(destinoArchivoCopiar):
+                    ventanaDialogo = QMessageBox()
+                    ventanaDialogo.setIcon(QMessageBox.Information)
+                    ventanaDialogo.setWindowTitle('Informacion')
+                    ventanaDialogo.setText("Ya has guardado una cancion con el mismo \n"
+                    "nombre, por tal motivo no se realizo la copia\n""de la cancion seleccionada ")
+                    ventanaDialogo.setStandardButtons(QMessageBox.Ok)
+                    btn_ok = ventanaDialogo.button(QMessageBox.Ok)
+                    btn_ok.setText('Entendido')
+                    ventanaDialogo.exec_()
+                else:
+                    shutil.copyfile(origenArchivoCopiar,destinoArchivoCopiar)
+                    self.senal_cancionAgregada.emit(soloNombreCancion)
             else:
                 print("La duracion del archivo no puede ser mayor a los 6 minutos")
                 ventanaDialogo = QMessageBox()
